refactored Data Canvas with component composition
