+++ src/main/components/DataTree.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import ReactJson from 'react-json-view';
 
 const styles = {
