+++ src/main/containers/DataCanvas.jsx
@@ -5,7 +5,7 @@ import EmptyState from '../components/EmptyState.jsx';
 
 const DataCanvas = (props) => {
   const {
-    data, treeCount, updateTreeCount, options,
+    data, treeCount, options,
   } = props;
 
   // Display empty state for no data
@@ -20,7 +20,6 @@ const DataCanvas = (props) => {
   }
   // Increment tree count and render data
   // TreeCount will always be 0 or 1 -- fix
-  if (!treeCount) updateTreeCount(1);
   return (
     <DataTree
       treeCount={treeCount}
