+++ src/main/containers/Panels.jsx
@@ -3,14 +3,13 @@ import styled from 'styled-components';
 import SourcePanel from './SourcePanel.jsx';
 import TestPanel from './TestPanel.jsx';
 import DestinationPanel from './DestinationPanel.jsx';
-// import { testsData } from '../dummyData';
-// import { smallData } from '../dummyData';
+import DataCanvas from './DataCanvas.jsx';
+// import { smallData, testsData } from '../dummyData';
 
 const Panels = () => {
   const [activePanel, setActivePanel] = useState('source');
-  const [treeCount, setTreeCount] = useState(0);
   const [data, setData] = useState(undefined);
-  const [testsDiff, setTestsDiff] = useState([ {} ]);
+  const [testsDiff, setTestsDiff] = useState([{}]);
   const [cursor, setCursor] = useState('default');
 
   const PanelsWrapper = styled.section`
@@ -18,36 +17,46 @@ const Panels = () => {
     height: 80vh;
     cursor: ${cursor};
   `;
-  // Tests are objects with
-  // { payload: JSON that represents test,
-  //   status: initial value of '' };
+
+  // Create DataCanvas component for component composition to
+  // be passed down as a prop to the panels that render it
+  const dataTreeOptions = {
+    onAdd: false,
+    onEdit: false,
+    onDelete: false,
+    enableClipboard: false,
+  };
+
+  const datacanvas = (
+    <DataCanvas
+      treeCount={'rawdata'}
+      data={data}
+      options={dataTreeOptions} />
+  );
 
   return (
     <PanelsWrapper>
-        <SourcePanel
-          onClickFunction={() => setActivePanel('source')}
-          treeCount={treeCount}
-          updateTreeCount={setTreeCount}
-          data={data}
-          setData={setData}
-          active={(activePanel === 'source')}
-          setCursor={setCursor} />
+      <SourcePanel
+        onClickFunction={() => setActivePanel('source')}
+        datacanvas={datacanvas}
+        setData={setData}
+        active={(activePanel === 'source')}
+        setCursor={setCursor} />
 
-        <TestPanel 
-          onClickFunction={() => setActivePanel('test')}
-          treeCount={treeCount}
-          updateTreeCount={setTreeCount}
-          data={data}
-          active={(activePanel === 'test')}
-          setCursor={setCursor}
-          testsDiff={testsDiff}
-          setTestsDiff={setTestsDiff} />
+      <TestPanel
+        onClickFunction={() => setActivePanel('test')}
+        datacanvas={datacanvas}
+        data={data}
+        active={(activePanel === 'test')}
+        setCursor={setCursor}
+        testsDiff={testsDiff}
+        setTestsDiff={setTestsDiff} />
 
-        <DestinationPanel 
-          onClickFunction={() => setActivePanel('dest')}
-          active={(activePanel === 'dest')}
-          setCursor={setCursor}
-          testsDiff={testsDiff} />
+      <DestinationPanel
+        onClickFunction={() => setActivePanel('dest')}
+        active={(activePanel === 'dest')}
+        setCursor={setCursor}
+        testsDiff={testsDiff} />
     </PanelsWrapper>
   );
 };
