+++ src/main/containers/SourcePanel.jsx
@@ -1,34 +1,19 @@
-import React, { useState, useContext } from 'react';
-import DataCanvas from './DataCanvas.jsx';
+import React from 'react';
 import RequestBar from '../components/RequestBar.jsx';
 import StyledPanel from './StyledPanel.jsx';
-import { TestsContext } from './../testsContext.js';
 
 
 const SourcePanel = (props) => {
   const {
-    treeCount, updateTreeCount, data, setData, active, onClickFunction, setCursor
+    setData, active, onClickFunction, setCursor
   } = props;
 
-  const [tests, setTests] = useContext(TestsContext);
-
-  const dataTreeOptions = {
-    onAdd: false,
-    onEdit: false,
-    onDelete: false,
-    enableClipboard: false,
-  };
 
   if (active) {
     return (
       <StyledPanel active={active} onMouseOver={() => setCursor('default')}>
         <RequestBar SourceOrDest='source' setData={setData} />
-        <DataCanvas
-          treeCount={treeCount}
-          updateTreeCount={updateTreeCount}
-          data={data}
-          options={dataTreeOptions}
-        />
+        {props.datacanvas}
       </StyledPanel>
     );
   }
@@ -37,7 +22,7 @@ const SourcePanel = (props) => {
     <StyledPanel
       onClick={onClickFunction}
       active={active}
-      onMouseOver={() => setCursor('pointer')} 
+      onMouseOver={() => setCursor('pointer')}
     >
       <h1>Source</h1>
     </StyledPanel>
