+++ src/main/containers/TestPanel.jsx
@@ -1,19 +1,17 @@
 import React, { useState, useContext } from 'react';
 // sample JSON to pass down as props. Will be able to remove as the project evolves.
 // import { largeData, smallData } from '../dummyData';
-import DataCanvas from './DataCanvas.jsx';
 import StyledPanel from './StyledPanel.jsx';
 import { TestsContext } from './../testsContext.js';
 import Button from '../components/Button.jsx';
 import DataTree from '../components/DataTree.jsx';
 
 // will need to get data from the get request to pass to the formatted view
-
 const TestPanel = (props) => {
   const {
-    active, updateTreeCount, data, onClickFunction, setTestsDiff, testsDiff, setCursor
+    active, datacanvas, data, onClickFunction, setTestsDiff, testsDiff, setCursor
   } = props;
-  
+
   const [tests, setTests] = useContext(TestsContext);
 
   const [testsListCounter, setTestsListCounter] = useState(0);
@@ -23,8 +21,8 @@ const TestPanel = (props) => {
     onDelete: true,
     enableClipboard: false,
   };
-  
-  function saveUpdatedTree(newData, arrayPosition, newValue, name, namespace) {
+
+  function saveUpdatedTree (newData, arrayPosition, newValue, name, namespace) {
     const testsClone = [...tests];
     const testsDiffClone = [...testsDiff];
     testsClone[arrayPosition].payload = newData;
@@ -54,12 +52,12 @@ const TestPanel = (props) => {
     i += 1;
   });
 
-  function createNewTest() {
+  function createNewTest () {
     const testsClone = [...tests];
     const testsDiffClone = [...testsDiff];
     testsClone.push({ payload: data, status: '' });
     testsDiffClone.push({});
-  
+
     // the ID of the test will be the same as the position in the array
     setTestsListCounter(testsListCounter + 1);
     setTests(testsClone);
@@ -68,19 +66,15 @@ const TestPanel = (props) => {
 
   if (active) {
     return (
-        <StyledPanel active={active} onMouseOver={() => setCursor('default')}>
-          <div>
-            {data && <h3>Server Response</h3>}
-            <DataCanvas
-              data={data}
-              updateTreeCount={updateTreeCount}
-              options={dataTreeOptions}
-            />
-            {data && <Button enabled={true} onClick={createNewTest}> New Test </Button>}
-          </div>
-          
-          {testsList}
-        </StyledPanel>
+      <StyledPanel active={active} onMouseOver={() => setCursor('default')}>
+        <div>
+          {data && <h3>Server Response</h3>}
+          {datacanvas}
+          {data && <Button enabled={true} onClick={createNewTest}> New Test </Button>}
+        </div>
+
+        {testsList}
+      </StyledPanel>
     )
   }
 
@@ -89,7 +83,7 @@ const TestPanel = (props) => {
       onClick={onClickFunction}
       active={active}
       onMouseOver={() => setCursor('pointer')} >
-        <h1>Test</h1>
+      <h1>Test</h1>
     </StyledPanel>
   )
 };
